+++ ruby/red-arrow/test/raw-records/test-dense-union-array.rb
@@ -94,7 +94,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:null, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_boolean
@@ -104,7 +104,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:boolean, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_int8
@@ -114,7 +114,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:int8, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_uint8
@@ -124,7 +124,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:uint8, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_int16
@@ -134,7 +134,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:int16, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_uint16
@@ -144,7 +144,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:uint16, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_int32
@@ -154,7 +154,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:int32, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_uint32
@@ -164,7 +164,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:uint32, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_int64
@@ -174,7 +174,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:int64, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_uint64
@@ -184,7 +184,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:uint64, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_float
@@ -194,7 +194,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:float, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_double
@@ -204,7 +204,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:double, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_binary
@@ -214,7 +214,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:binary, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_string
@@ -224,7 +224,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:string, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_date32
@@ -234,7 +234,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:date32, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_date64
@@ -244,7 +244,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:date64, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_timestamp_second
@@ -258,7 +258,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_timestamp_milli
@@ -272,7 +272,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_timestamp_micro
@@ -286,7 +286,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_timestamp_nano
@@ -300,7 +300,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_time32_second
@@ -316,7 +316,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_time32_milli
@@ -332,7 +332,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_time64_micro
@@ -348,7 +348,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_time64_nano
@@ -364,7 +364,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_decimal128
@@ -379,7 +379,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_decimal256
@@ -394,7 +394,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_month_interval
@@ -404,7 +404,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:month_interval, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_day_time_interval
@@ -414,7 +414,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:day_time_interval, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_month_day_nano_interval
@@ -424,7 +424,7 @@ module RawRecordsDenseUnionArrayTests
     ]
     target = build(:month_day_nano_interval, records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_list
@@ -441,7 +441,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_struct
@@ -461,7 +461,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_map
@@ -476,7 +476,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_sparse_union
@@ -502,7 +502,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(remove_field_names(records)),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_dense_union
@@ -528,7 +528,7 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(remove_field_names(records)),
-                 target.raw_records)
+                 actual_records(target))
   end
 
   def test_dictionary
@@ -545,7 +545,31 @@ module RawRecordsDenseUnionArrayTests
                    },
                    records)
     assert_equal(remove_field_names(records),
-                 target.raw_records)
+                 actual_records(target))
+  end
+end
+
+class EachRawRecordRecordBatchDenseUnionArrayTest < Test::Unit::TestCase
+  include RawRecordsDenseUnionArrayTests
+
+  def build(type, records)
+    build_record_batch(type, records)
+  end
+
+  def actual_records(target)
+    target.each_raw_record.to_a
+  end
+end
+
+class EachRawRecordTableDenseUnionArrayTest < Test::Unit::TestCase
+  include RawRecordsDenseUnionArrayTests
+
+  def build(type, records)
+    build_record_batch(type, records).to_table
+  end
+
+  def actual_records(target)
+    target.each_raw_record.to_a
   end
 end
 
@@ -555,6 +579,10 @@ class RawRecordsRecordBatchDenseUnionArrayTest < Test::Unit::TestCase
   def build(type, records)
     build_record_batch(type, records)
   end
+
+  def actual_records(target)
+    target.raw_records
+  end
 end
 
 class RawRecordsTableDenseUnionArrayTest < Test::Unit::TestCase
@@ -562,5 +590,9 @@ class RawRecordsTableDenseUnionArrayTest < Test::Unit::TestCase
 
   def build(type, records)
     build_record_batch(type, records).to_table
+  end
+
+  def actual_records(target)
+    target.raw_records
   end
 end
